requirements.txt
appdirs==1.4.3
appnope==0.1.0
aspy.yaml==1.2.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
black==19.3b0
bs4==0.0.1
certifi==2019.3.9
cfgv==1.6.0
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.1
django-extensions==2.1.6
django-localflavor==2.1
django-polymorphic==2.0.3
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.9.2
Faker==1.0.5
identify==1.4.1
idna==2.8
importlib-metadata==0.9
importlib-resources==1.0.2
ipdb==0.12
ipython==7.4.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
jellyfish==0.5.6
Jinja2==2.10.1
jsonfield==2.0.2
MarkupSafe==1.1.1
nodeenv==1.3.3
numpy==1.16.2
oauthlib==3.0.1
openapi-codec==1.3.2
pandas==0.24.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pre-commit==1.15.2
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyarrow==0.13.0
Pygments==2.3.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
text-unidecode==1.2
toml==0.10.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.2
us==1.0.0
virtualenv==16.4.3
wcwidth==0.1.7
Werkzeug==0.15.2
xlrd==1.2.0
zipp==0.3.3
